ApiPythonRNA/requirements.txt
fastapi
uvicorn[standard]
spacy
sqlalchemy
pymysql
python-dotenv
requests